Silence Linux compile warnings
- When compiled on Linux, the compiler reports a couple warnings

- Cast "rank" to "size_t" in `assign_subdomain()`
- Initialize "spatial1D" before use in `read_spatial_topo_climate_site_inputs()`

+++ src/SW_Domain.c
@@ -145,7 +145,8 @@ static void assign_subdomain(
     size_t domStartIndexProg[],
     size_t domCountsProg[]
 ) {
-    const size_t chunkYX = (isSimDomDiscrete) ? rank : rank / nChunks[1];
+    const size_t chunkYX =
+        (isSimDomDiscrete) ? (size_t) rank : rank / nChunks[1];
     const size_t chunkX = (isSimDomDiscrete) ? 0 : rank % nChunks[1];
 
     domStartIndexProg[0] = startY[chunkYX];
